# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/files/signature": {
      "get": {
        "description": "Generates a Cloudinary signature for client-side uploads with optional folder and public ID.",
        "operationId": "FilesController_getCloudinarySignature",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSignatureDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cloudinary signature generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "signature": "abc123def456ghi789",
                    "timestamp": 1704067200,
                    "cloudName": "your-cloud-name",
                    "apiKey": "your-api-key",
                    "folder": "lume_ecom_uploads"
                  }
                }
              }
            }
          }
        },
        "summary": "Generate Cloudinary signature",
        "tags": [
          "Files"
        ]
      }
    },
    "/files/upload": {
      "post": {
        "description": "Uploads a single file to Cloudinary with optional transformations and folder organization.",
        "operationId": "FilesController_uploadFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "File upload with optional parameters",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "File to upload"
                  },
                  "folder": {
                    "type": "string",
                    "description": "Cloudinary folder path",
                    "example": "products"
                  },
                  "publicId": {
                    "type": "string",
                    "description": "Public ID for the file",
                    "example": "product_123_image"
                  },
                  "resourceType": {
                    "type": "string",
                    "enum": [
                      "image",
                      "video",
                      "raw",
                      "auto"
                    ],
                    "description": "Resource type",
                    "example": "image"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "public_id": "lume_ecom_uploads/products/abc123",
                    "url": "https://res.cloudinary.com/lume/image/upload/v1729990123/lume_ecom_uploads/products/abc123.webp",
                    "format": "webp",
                    "bytes": 245231,
                    "width": 800,
                    "height": 600
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid file or upload parameters",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "message": "Invalid file format or size"
                  }
                }
              }
            }
          }
        },
        "summary": "Upload file to Cloudinary",
        "tags": [
          "Files"
        ]
      }
    },
    "/files/upload-multiple": {
      "post": {
        "description": "Uploads multiple files to Cloudinary in batch with optional transformations.",
        "operationId": "FilesController_uploadMultipleFiles",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Multiple files upload",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Files to upload"
                  },
                  "folder": {
                    "type": "string",
                    "description": "Cloudinary folder path",
                    "example": "products"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Files uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "files": [
                      {
                        "success": true,
                        "public_id": "lume_ecom_uploads/products/file1",
                        "url": "https://res.cloudinary.com/lume/image/upload/v1729990123/lume_ecom_uploads/products/file1.webp",
                        "format": "webp",
                        "bytes": 245231,
                        "width": 800,
                        "height": 600
                      },
                      {
                        "success": true,
                        "public_id": "lume_ecom_uploads/products/file2",
                        "url": "https://res.cloudinary.com/lume/image/upload/v1729990124/lume_ecom_uploads/products/file2.webp",
                        "format": "webp",
                        "bytes": 198765,
                        "width": 800,
                        "height": 600
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "summary": "Upload multiple files to Cloudinary",
        "tags": [
          "Files"
        ]
      }
    },
    "/files/{publicId}": {
      "delete": {
        "description": "Deletes a file from Cloudinary using its public ID.",
        "operationId": "FilesController_deleteFile",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteFileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "result": "ok"
                  }
                }
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "message": "File not found"
                  }
                }
              }
            }
          }
        },
        "summary": "Delete file from Cloudinary",
        "tags": [
          "Files"
        ]
      },
      "get": {
        "description": "Retrieves detailed information about a file stored in Cloudinary.",
        "operationId": "FilesController_getFileInfo",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetFileInfoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "public_id": "products/product_123_image",
                    "format": "jpg",
                    "width": 800,
                    "height": 600,
                    "bytes": 1024000,
                    "url": "https://res.cloudinary.com/your-cloud/image/upload/v1234567890/products/product_123_image.jpg",
                    "secure_url": "https://res.cloudinary.com/your-cloud/image/upload/v1234567890/products/product_123_image.jpg",
                    "created_at": "2024-01-15T10:30:00Z"
                  }
                }
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": false,
                    "message": "File not found"
                  }
                }
              }
            }
          }
        },
        "summary": "Get file information",
        "tags": [
          "Files"
        ]
      }
    },
    "/files/generate-url": {
      "post": {
        "description": "Generates an optimized image URL with transformations for better performance (default: 600x600, fill crop, auto format/webp).",
        "operationId": "FilesController_generateOptimizedUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateUrlDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Optimized URL generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "optimizedUrl": "https://res.cloudinary.com/your-cloud/image/upload/w_600,h_600,c_fill,q_auto,f_auto/products/product_123_image.webp"
                  }
                }
              }
            }
          }
        },
        "summary": "Generate optimized image URL",
        "tags": [
          "Files"
        ]
      }
    },
    "/files/thumbnail/{publicId}": {
      "get": {
        "description": "Generates a thumbnail URL for product images with optimal sizing.",
        "operationId": "FilesController_generateThumbnailUrl",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thumbnail URL generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "url": "https://res.cloudinary.com/your-cloud/image/upload/w_300,h_300,c_fill,q_auto,f_auto/products/product_123_image.jpg"
                  }
                }
              }
            }
          }
        },
        "summary": "Generate thumbnail URL",
        "tags": [
          "Files"
        ]
      }
    },
    "/files/gallery/{publicId}": {
      "get": {
        "description": "Generates a gallery URL for product images with optimal sizing for galleries.",
        "operationId": "FilesController_generateGalleryUrl",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Gallery URL generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "url": "https://res.cloudinary.com/your-cloud/image/upload/w_800,h_600,c_fill,q_auto,f_auto/products/product_123_image.jpg"
                  }
                }
              }
            }
          }
        },
        "summary": "Generate gallery URL",
        "tags": [
          "Files"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ecom API",
    "description": "E-commerce REST API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GenerateSignatureDto": {
        "type": "object",
        "properties": {
          "folder": {
            "type": "string",
            "description": "Cloudinary folder path",
            "example": "products"
          },
          "publicId": {
            "type": "string",
            "description": "Public ID for the file",
            "example": "product_123_image"
          }
        }
      },
      "DeleteFileDto": {
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string",
            "description": "Public ID of the file to delete",
            "example": "products/product_123_image"
          },
          "resourceType": {
            "type": "string",
            "description": "Resource type of the file",
            "example": "image",
            "enum": [
              "image",
              "video",
              "raw"
            ]
          }
        },
        "required": [
          "publicId"
        ]
      },
      "GetFileInfoDto": {
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string",
            "description": "Public ID of the file",
            "example": "products/product_123_image"
          },
          "resourceType": {
            "type": "string",
            "description": "Resource type of the file",
            "example": "image",
            "enum": [
              "image",
              "video",
              "raw"
            ]
          }
        },
        "required": [
          "publicId"
        ]
      },
      "GenerateUrlDto": {
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string",
            "description": "Public ID of the image",
            "example": "products/product_123_image"
          },
          "width": {
            "type": "number",
            "description": "Image width",
            "example": 800
          },
          "height": {
            "type": "number",
            "description": "Image height",
            "example": 600
          },
          "crop": {
            "type": "string",
            "description": "Crop mode",
            "example": "fill"
          },
          "quality": {
            "type": "string",
            "description": "Image quality",
            "example": "auto"
          },
          "format": {
            "type": "string",
            "description": "Image format",
            "example": "auto"
          }
        },
        "required": [
          "publicId"
        ]
      },
...
    }
  }
}